Update RequestHandler.cs
When you try to connect a websocket which has a diffrent port number than 80, request thinks that this is a https connection and throw exception. Because of this can't open https://web.whatsapp.com/ or can't connect to ws://188.166.50.126:8080. With this change it is working. You can try to connect this web socket on this site http://www.websocket.org/echo.html . A normal websocket work with older code but diffrent port like 8080 not working with old code.